CRUD for semester only for data of user who is logged in

+++ NOTEit/NOTEit/Controllers/SemesterController.cs
@@ -8,13 +8,14 @@ using NOTEit.ViewModels.Semester;
 
 namespace NOTEit.Controllers
 {
+    [Authorize]
     public class SemesterController : Controller
     {
         private readonly ApplicationDbContext _db = new ApplicationDbContext();
 
         public ActionResult Index()
         {
-            return View(_db.Semesters.ToList());
+            return View(_db.Semesters.Where(x => x.Owner.Id == User.Identity.GetUserId()).ToList());
         }
 
         public ActionResult Details(int? id)
@@ -24,7 +25,7 @@ namespace NOTEit.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             var semester = _db.Semesters.Find(id);
-            if (semester == null)
+            if (semester == null || semester.Owner.Id != User.Identity.GetUserId())
             {
                 return HttpNotFound();
             }
@@ -36,7 +37,7 @@ namespace NOTEit.Controllers
             return View(
                 new SemesterFormViewModel
                 {
-                    AllSubjects = _db.Subjects.ToList()
+                    AllSubjects = _db.Subjects.Where(x => x.Semesters.Any(y => y.Owner.Id == User.Identity.GetUserId())).ToList()
                 }
             );
         }
@@ -77,7 +78,7 @@ namespace NOTEit.Controllers
                     Id = semester.Id,
                     Name = semester.Name,
                     Subjects = semester.Subjects.Select(x => x.Id).ToList(),
-                    AllSubjects = _db.Subjects.ToList()
+                    AllSubjects = _db.Subjects.Where(x => x.Semesters.Any(y => y.Owner.Id == User.Identity.GetUserId())).ToList()
                 }
             );
         }
@@ -105,7 +106,7 @@ namespace NOTEit.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             var semester = _db.Semesters.Find(id);
-            if (semester == null)
+            if (semester == null || semester.Owner.Id != User.Identity.GetUserId())
             {
                 return HttpNotFound();
             }
@@ -117,7 +118,7 @@ namespace NOTEit.Controllers
         public ActionResult Delete(int id)
         {
             var semester = _db.Semesters.Find(id);
-            if (semester == null) return View("Error");
+            if (semester == null || semester.Owner.Id != User.Identity.GetUserId()) return View("Error");
             _db.Semesters.Remove(semester);
             _db.SaveChanges();
             return RedirectToAction("Index");
